requirements.txt
Django
djangorestframework
python-docx
python-pptx
PyPDF2
pdf2image
Pillow
# pdf-lib
pymupdf
reportlab
python-magic
celery
redis